Split out std err listener
