+++ src/main/kotlin/com/dprint/services/editorservice/v4/EditorServiceV4.kt
@@ -2,9 +2,9 @@ package com.dprint.services.editorservice.v4
 
 import com.dprint.config.ProjectConfiguration
 import com.dprint.i18n.DprintBundle
-import com.dprint.services.editorservice.EditorProcess
 import com.dprint.services.editorservice.FormatResult
 import com.dprint.services.editorservice.IEditorService
+import com.dprint.services.editorservice.process.EditorProcess
 import com.dprint.utils.infoLogWithConsole
 import com.dprint.utils.warnLogWithConsole
 import com.intellij.openapi.components.Service
